leerie: Consolidate four duplicated _entry_sh_rootful_block extractors
test_claude_json_dir_mount.py, test_claude_json_mount.py,
test_home_leerie_ownership.py, and test_tmp_cache_writable.py each
defined a byte-identical _entry_sh_rootful_block(). One docstring
already claimed the block was "deliberately reused" across these
files, but each was a separate copy. Moved the single definition into
tests/stub_helpers.py and imported it from all four call sites.
