+++ tests/test_tmp_cache_writable.py
@@ -24,6 +24,8 @@ from __future__ import annotations
 
 from pathlib import Path
 
+from tests.stub_helpers import _entry_sh_rootful_block
+
 REPO_ROOT = Path(__file__).resolve().parent.parent
 DOCKERFILE = REPO_ROOT / "Dockerfile"
 ENTRY_SH = REPO_ROOT / "scripts" / "container-entry.sh"
@@ -34,14 +36,6 @@ def _dockerfile_tmp_cache_block() -> str:
     marker = "RUN chown -R leerie: /tmp/.cache"
     start = text.index(marker)
     end = text.index("\n\n", start)
-    return text[start:end]
-
-
-def _entry_sh_rootful_block() -> str:
-    text = ENTRY_SH.read_text()
-    marker = 'if [ "$ROOTLESS" != "true" ] && getent passwd leerie'
-    start = text.index(marker)
-    end = text.index("\nfi", start)
     return text[start:end]
